ai-backend/requirements.txt
flask
flask-cors
python-dotenv
PyPDF2
ibm-watsonx-ai
waitress
